third_party/blink/web_tests/wpt_internal/display-lock/rendersubtree/activation/activate-update-same-frame.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Display Locking: activate update same frame</title>
<link rel="author" title="Rakina Zata Amni" href="mailto:rakina@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/utils.js"></script>

<style>
.spacer {
  width: 150px;
  height: 3000px;
  background: lightblue;
}
#container {
  width: 150px;
  height: 150px;
  background: lightgreen;
}
#target {
  width: 100px;
  height: 100px;
  background: green;
}
</style>

<div class="spacer"></div>
<div id="container"><div id="target"></div></div>

<script>
async_test((t) => {
  async function runTest() {
    let container = document.getElementById("container");

    let acquire_promise = setInvisibleActivatable(container);
    await acquire_promise;

    target.scrollIntoView();
    let update_promise = container.updateRendering();
    await update_promise;
    t.step(() => assert_equals(container.renderSubtree, "", "context after update & activation is unlocked"));
    t.done();
  }

  window.onload = function() {
    requestAnimationFrame(() => requestAnimationFrame(runTest));
  };
}, "scrollIntoView and updating on the same frame should work");

</script>
</html>
